By popular request, here is an updated walkthrough for TSROC
This walkthrough contains spoilers as far as telling you where to go and what to do for the first several hours of the game.

Read on if you prefer a guided experience, but know there is no wrong way to play. While the game has an intended path, it also has numerous optional routes and a fairly open-world structure.